feat: add just one persisted field

+++ api/src/main/java/mz/co/checkmob/api/jobs/domain/RequestExecutorCommand.java
@@ -5,6 +5,6 @@ import lombok.Data;
 @Data
 public class RequestExecutorCommand {
     private RequestFrequency frequency;
-    private Integer every;
+    private Long every;
     private TimeUnity unity;
 }
